#14 [chore] LoginResponse 패키지 변경
