--- data/src/main/java/hous/release/data/model/response/LoginResponse.kt
@@ -1,15 +1,0 @@
-package hous.release.data.model.response
-
-import com.google.gson.annotations.SerializedName
-
-data class LoginResponse(
-    val token: Token,
-    val userId: String
-)
-
-data class Token(
-    @SerializedName("access_token")
-    val accessToken: String,
-    @SerializedName("refresh_token")
-    val refreshToken: String
-)
